Move app_user and db creation into the postgresql role
This is to fix issues when bringing up a cluster for the first time. We
couldn't setup the app first because postgresql wasn't configured yet
and we couldn't configure postgresql first because it needed the
app_user to be created first. This moves all postgresql related tasks
into the postgresql role to fix that.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>