fix intermittent 504 when requesting the latest for a repo
This resolves a situation where using the 'latest' sha1 feature in the repos
endpoint would result in a 504 even if there are repos built for that project/ref.

For example, the following url would give a 504 intermittently even though there
are many builds for the ceph master branch for xenial:
https://shaman.ceph.com/api/repos/ceph/master/latest/ubuntu/xenial/repo/

This scenario would occur when:

- a repo for ceph/master/sah1Y was built and completed for xenial
- a repo for ceph/master/sha1X was built and completed for centos7
- a repo for ceph/master/sha1X is still being built for xenial
- a request to the repo endpoint for ceph/master/latest/ubuntu/xenial
would select sha1x instead of sha1Y which would *actually* be the latest
sha1 that is ready for use.

This patch fixes that by having the sha1 contoller filter by distro and
distro_version if sha1=latest is used.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>