adds a new endpoint that redirects to the repo contents
This new endpoint will allow you to construct a yum or apt repo file
that will always redirect to the latest version of a repo. For example
the following url:

   shaman.ceph.com/api/repos/ceph/master/latest/centos/7/x86_64/noarch/

Would redirect to the noarch directory for the latest centos 7 x86_64
build of ceph master.

If you only include the architecture, it would link to the root of the
repo directory.

This also works at the 'flavors' endpoint, so if you need the latest
notcmalloc ceph master build then this url would work:

    shaman.ceph.com/api/repos/ceph/master/latest/centos/7/flavors/notcmalloc/x86_64/

Signed-off-by: Andrew Schoen <aschoen@redhat.com>