requirements: Pin versions
A minor update to shaman was made.  When it was redeployed, a newer version of pika got installed that relies on a version of enum that is only available in python3 environements.  enum was backported as enum34 but has to be pulled in manually.

So we're pinning all dependencies until shaman gets updated to be python3-compatible.

Signed-off-by: David Galloway <dgallowa@redhat.com>