deploy: remove hot-standby server
When updating the hosts, the differences between configuring
hot-standby on postgresql 9.5 -> 16 are just too great.
I think the hot-standby feature was of questionable value
anyway, and it super-complicates the update procedure, and
shaman is due for some major architectural changes soon
anyway.

If this horribly affects performance we can circle back,
but I very much doubt it will.

Signed-off-by: Dan Mick <dan.mick@redhat.com>